magpie_agent/pyproject.toml
[project]
name = "magpie-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.14.0",
    "httpx[socks]>=0.28.1",
    "mcp>=1.26.0",
    "openai>=2.23.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
